Move domain-specific entities to the appropriate files
