+++ homeassistant/components/starlink/binary_sensor.py
@@ -2,7 +2,16 @@
 
 from __future__ import annotations
 
-from homeassistant.components.binary_sensor import BinarySensorDeviceClass
+from collections.abc import Callable
+from dataclasses import dataclass
+
+from starlink_grpc import AlertDict, ObstructionDict, StatusDict
+
+from homeassistant.components.binary_sensor import (
+    BinarySensorDeviceClass,
+    BinarySensorEntity,
+    BinarySensorEntityDescription,
+)
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity import EntityCategory
@@ -10,7 +19,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .const import DOMAIN
 from .coordinator import StarlinkUpdateCoordinator
-from .entity import StarlinkBinarySensorEntity, StarlinkBinarySensorEntityDescription
+from .entity import StarlinkEntity
 
 # This stores device_class overrides for alert types that we know aren't "problems".
 DEVICE_CLASS_OVERRIDES = {
@@ -60,3 +69,38 @@ async def async_setup_entry(
     )
 
     async_add_entities(entities)
+
+
+@dataclass
+class StarlinkBinarySensorEntityDescriptionMixin:
+    """Mixin for required keys."""
+
+    value_fn: Callable[[tuple[StatusDict, ObstructionDict, AlertDict]], bool | None]
+
+
+@dataclass
+class StarlinkBinarySensorEntityDescription(
+    BinarySensorEntityDescription, StarlinkBinarySensorEntityDescriptionMixin
+):
+    """Describes a Starlink binary sensor entity."""
+
+
+class StarlinkBinarySensorEntity(StarlinkEntity, BinarySensorEntity):
+    """A BinarySensorEntity for Starlink devices. Handles creating unique IDs."""
+
+    entity_description: StarlinkBinarySensorEntityDescription
+
+    def __init__(
+        self,
+        coordinator: StarlinkUpdateCoordinator,
+        description: StarlinkBinarySensorEntityDescription,
+    ) -> None:
+        """Initialize the binary sensor."""
+        super().__init__(coordinator)
+        self.entity_description = description
+        self._attr_unique_id = f"{self.coordinator.data[0]['id']}_{description.key}"
+
+    @property
+    def is_on(self) -> bool | None:
+        """Calculate the dinary sensor value from the entity description."""
+        return self.entity_description.value_fn(self.coordinator.data)
