+++ homeassistant/components/starlink/sensor.py
@@ -1,17 +1,77 @@
 """Contains sensors exposed by the Starlink integration."""
 from __future__ import annotations
 
+from collections.abc import Callable
+from dataclasses import dataclass
 from datetime import datetime, timedelta
 
-from homeassistant.components.sensor import SensorDeviceClass, SensorStateClass
+from starlink_grpc import AlertDict, ObstructionDict, StatusDict
+
+from homeassistant.components.sensor import (
+    SensorDeviceClass,
+    SensorEntity,
+    SensorEntityDescription,
+    SensorStateClass,
+)
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import DEGREE, UnitOfDataRate, UnitOfTime
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity import EntityCategory
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
+from homeassistant.helpers.typing import StateType
 
 from .const import DOMAIN
-from .entity import StarlinkSensorEntity, StarlinkSensorEntityDescription
+from .coordinator import StarlinkUpdateCoordinator
+from .entity import StarlinkEntity
+
+
+async def async_setup_entry(
+    hass: HomeAssistant, entry: ConfigEntry, async_add_entities: AddEntitiesCallback
+) -> None:
+    """Set up all sensors for this entry."""
+    coordinator = hass.data[DOMAIN][entry.entry_id]
+
+    async_add_entities(
+        StarlinkSensorEntity(coordinator, description) for description in SENSORS
+    )
+
+
+@dataclass
+class StarlinkSensorEntityDescriptionMixin:
+    """Mixin for required keys."""
+
+    value_fn: Callable[
+        [tuple[StatusDict, ObstructionDict, AlertDict]], datetime | StateType
+    ]
+
+
+@dataclass
+class StarlinkSensorEntityDescription(
+    SensorEntityDescription, StarlinkSensorEntityDescriptionMixin
+):
+    """Describes a Starlink sensor entity."""
+
+
+class StarlinkSensorEntity(StarlinkEntity, SensorEntity):
+    """A SensorEntity for Starlink devices. Handles creating unique IDs."""
+
+    entity_description: StarlinkSensorEntityDescription
+
+    def __init__(
+        self,
+        coordinator: StarlinkUpdateCoordinator,
+        description: StarlinkSensorEntityDescription,
+    ) -> None:
+        """Initialize the sensor."""
+        super().__init__(coordinator)
+        self.entity_description = description
+        self._attr_unique_id = f"{self.coordinator.data[0]['id']}_{description.key}"
+
+    @property
+    def native_value(self) -> StateType | datetime:
+        """Calculate the sensor value from the entity description."""
+        return self.entity_description.value_fn(self.coordinator.data)
+
 
 SENSORS: tuple[StarlinkSensorEntityDescription, ...] = (
     StarlinkSensorEntityDescription(
@@ -66,14 +126,3 @@ SENSORS: tuple[StarlinkSensorEntityDescription, ...] = (
         - timedelta(seconds=data[0]["uptime"]),
     ),
 )
-
-
-async def async_setup_entry(
-    hass: HomeAssistant, entry: ConfigEntry, async_add_entities: AddEntitiesCallback
-) -> None:
-    """Set up all sensors for this entry."""
-    coordinator = hass.data[DOMAIN][entry.entry_id]
-
-    async_add_entities(
-        StarlinkSensorEntity(coordinator, description) for description in SENSORS
-    )
